% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main_function.R
\name{applyThreshold}
\alias{applyThreshold}
\title{Apply Thresholding to VAR Coefficients}
\usage{
applyThreshold(a_mat, nr, nc, p, type = "soft")
}
\arguments{
\item{a_mat}{Numeric matrix. The coefficient matrix to be thresholded.}

\item{nr}{Integer. The number of rows in the original data.}

\item{nc}{Integer. The number of variables (columns) in the original data.}

\item{p}{Integer. The order of the VAR model.}

\item{type}{Character. The type of threshold to apply; either \code{"soft"} (default)
or \code{"hard"}.}
}
\value{
The thresholded coefficient matrix.
}
\description{
Applies a thresholding rule to a coefficient matrix by setting entries below a
certain threshold to zero. Two types of thresholding are available: "soft" and "hard".
}
\keyword{internal}
