% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main_function.R
\name{computeResiduals}
\alias{computeResiduals}
\title{Compute VAR Model Residuals}
\usage{
computeResiduals(data, A)
}
\arguments{
\item{data}{A numeric matrix of the original time series (observations in rows).}

\item{A}{List. A list of VAR coefficient matrices (one for each lag).}
}
\value{
A numeric matrix of residuals.
}
\description{
Computes the residuals from a VAR model by subtracting the fitted values (obtained
from the estimated coefficient matrices) from the original time series data.
}
\keyword{internal}
