% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main_function.R
\name{duplicateMatrix}
\alias{duplicateMatrix}
\title{Construct Lagged Design Matrix for VAR}
\usage{
duplicateMatrix(data, p)
}
\arguments{
\item{data}{A numeric matrix with time series data (observations in rows).}

\item{p}{Integer. The order of the VAR model (number of lags).}
}
\value{
A numeric matrix with duplicated columns corresponding to lagged observations.
}
\description{
Duplicates the original data matrix to create a lagged predictor matrix for VAR estimation.
}
\keyword{internal}
