% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main_function.R
\name{estimateCovariance}
\alias{estimateCovariance}
\title{Estimate Covariance Matrix from Residuals}
\usage{
estimateCovariance(res, ...)
}
\arguments{
\item{res}{A numeric matrix of residuals from the VAR model.}

\item{...}{Additional arguments passed to \code{corpcor::cov.shrink} (if any).}
}
\value{
A numeric covariance matrix.
}
\description{
Estimates the covariance (or variance) matrix of the residuals using shrinkage estimation.
This function utilizes \code{corpcor::cov.shrink} for covariance estimation.
}
\keyword{internal}
