% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main_function.R
\name{splitMatrix}
\alias{splitMatrix}
\title{Split Coefficient Matrix into VAR Lags}
\usage{
splitMatrix(M, p)
}
\arguments{
\item{M}{A numeric matrix of coefficients.}

\item{p}{Integer. The order of the VAR model (number of lags).}
}
\value{
A list of \code{p} matrices, each of dimension (number of variables) x (number of variables).
}
\description{
Splits a matrix of estimated coefficients into a list of matrices,
each corresponding to one lag of the VAR model.
}
\keyword{internal}
